test(rdf-core): the cover law's refusal vectors live where the law lives
The format-neutral refusal pairs — the cover, the declared and
undeclared overlaps, the shared-cut chain, the span-shape refusals, the
digest, the empty document — move (not copy) from the Markdown codec's
integration suite into cover.rs's own test module, so a future codec
crate and a kernel-only refactor both get the signal from this crate's
own suite. The graph-level vectors, spans extracted from parsed
triples, stay with the codec that emits them.
