+++ crates/rdf-core/src/cover.rs
@@ -375,3 +375,237 @@ pub fn reconstruct(
         valid_up_to: e.utf8_error().valid_up_to(),
     })
 }
+
+#[cfg(test)]
+mod tests {
+    //! The refusal vectors of the cover law, each executed beside the
+    //! neighbouring input that still succeeds — the workspace refusal
+    //! discipline, applied where the law lives so a future codec crate
+    //! and a kernel-only refactor both get the signal from this crate's
+    //! own suite. The graph-level vectors — spans extracted from parsed
+    //! triples — stay with the Markdown codec that emits them.
+
+    use super::*;
+
+    /// A lawful cover of `"ab\ncd"`: two spans meeting at byte 2.
+    fn covered() -> (u64, ContentDigest, [VerbatimSpan<'static>; 2]) {
+        let text = "ab\ncd";
+        (
+            text.len() as u64,
+            ContentDigest::of(text.as_bytes()),
+            [
+                VerbatimSpan {
+                    byte_start: 0,
+                    byte_end: 2,
+                    text: "ab",
+                    continues: None,
+                },
+                VerbatimSpan {
+                    byte_start: 2,
+                    byte_end: 5,
+                    text: "\ncd",
+                    continues: None,
+                },
+            ],
+        )
+    }
+
+    #[test]
+    fn a_full_cover_decodes_and_a_missing_span_names_the_uncovered_run() {
+        let (len, digest, spans) = covered();
+        assert_eq!(reconstruct(len, &digest, &spans).as_deref(), Ok("ab\ncd"));
+        assert_eq!(
+            reconstruct(len, &digest, &spans[..1]),
+            Err(ReconstructError::UncoveredRange {
+                byte_start: 2,
+                byte_end: 5,
+            })
+        );
+        assert_eq!(
+            reconstruct(len, &digest, &spans[1..]),
+            Err(ReconstructError::UncoveredRange {
+                byte_start: 0,
+                byte_end: 2,
+            })
+        );
+        // A declared length past the cover is the uncovered run it is —
+        // found before any buffer exists, so a graph cannot buy an
+        // allocation with a number.
+        assert_eq!(
+            reconstruct(u64::MAX, &digest, &spans),
+            Err(ReconstructError::UncoveredRange {
+                byte_start: 5,
+                byte_end: u64::MAX,
+            })
+        );
+    }
+
+    #[test]
+    fn an_overlap_is_refused_undeclared_and_admitted_declared_even_though_the_bytes_agree_either_way()
+     {
+        let text = "abcdef";
+        let len = text.len() as u64;
+        let digest = ContentDigest::of(text.as_bytes());
+        let piece = VerbatimSpan {
+            byte_start: 0,
+            byte_end: 4,
+            text: "abcd",
+            continues: None,
+        };
+        let agreeing = |continues| VerbatimSpan {
+            byte_start: 2,
+            byte_end: 6,
+            text: "cdef",
+            continues,
+        };
+        // Undeclared: the shared bytes agree — they came from the same
+        // file — and that is exactly why agreement alone must admit
+        // nothing.
+        assert_eq!(
+            reconstruct(len, &digest, &[piece, agreeing(None)]),
+            Err(ReconstructError::UndeclaredOverlap {
+                byte_start: 2,
+                byte_end: 6,
+                covered_start: 0,
+                covered_end: 4,
+            })
+        );
+        // A declaration naming a span the graph does not carry answers for
+        // nothing.
+        assert_eq!(
+            reconstruct(len, &digest, &[piece, agreeing(Some((0, 3)))]),
+            Err(ReconstructError::UndeclaredOverlap {
+                byte_start: 2,
+                byte_end: 6,
+                covered_start: 0,
+                covered_end: 4,
+            })
+        );
+        // The neighbour: the same overlap, declared to the piece that
+        // covers it, decodes.
+        assert_eq!(
+            reconstruct(len, &digest, &[piece, agreeing(Some((0, 4)))]).as_deref(),
+            Ok("abcdef")
+        );
+        // And a declared continuation that disagrees about a shared byte
+        // is named at that byte.
+        let disagreeing = VerbatimSpan {
+            byte_start: 2,
+            byte_end: 6,
+            text: "cXef",
+            continues: Some((0, 4)),
+        };
+        assert_eq!(
+            reconstruct(len, &digest, &[piece, disagreeing]),
+            Err(ReconstructError::OverlapDisagreement { at: 3 })
+        );
+    }
+
+    #[test]
+    fn a_chain_whose_pieces_share_a_cut_decodes_by_its_declarations() {
+        // The geometry the write-side check caught on a real split: two
+        // pieces end at one cut, and the continuation of the later one
+        // overlaps bytes the sorted walk credits to the earlier. The
+        // declaration — not the walk's frontier — is what answers.
+        let text = "ab\ncd\nef";
+        let len = text.len() as u64;
+        let digest = ContentDigest::of(text.as_bytes());
+        let spans = [
+            VerbatimSpan {
+                byte_start: 0,
+                byte_end: 2,
+                text: "ab",
+                continues: None,
+            },
+            VerbatimSpan {
+                byte_start: 1,
+                byte_end: 5,
+                text: "b\ncd",
+                continues: Some((0, 2)),
+            },
+            VerbatimSpan {
+                byte_start: 3,
+                byte_end: 5,
+                text: "cd",
+                continues: Some((1, 5)),
+            },
+            VerbatimSpan {
+                byte_start: 4,
+                byte_end: 8,
+                text: "d\nef",
+                continues: Some((3, 5)),
+            },
+        ];
+        assert_eq!(reconstruct(len, &digest, &spans).as_deref(), Ok(text));
+    }
+
+    #[test]
+    fn a_span_answers_for_its_own_shape_before_the_cover_is_asked() {
+        let (len, digest, [a, b]) = covered();
+        let mut inverted = a;
+        inverted.byte_start = 2;
+        inverted.byte_end = 0;
+        assert_eq!(
+            reconstruct(len, &digest, &[inverted, b]),
+            Err(ReconstructError::InvertedSpan {
+                byte_start: 2,
+                byte_end: 0,
+            })
+        );
+        let mut past = b;
+        past.byte_end = 6;
+        assert_eq!(
+            reconstruct(len, &digest, &[a, past]),
+            Err(ReconstructError::SpanOutOfBounds {
+                byte_start: 2,
+                byte_end: 6,
+                byte_length: 5,
+            })
+        );
+        let mut short = b;
+        short.text = "\ncd extra";
+        assert_eq!(
+            reconstruct(len, &digest, &[a, short]),
+            Err(ReconstructError::SpanLengthMismatch {
+                byte_start: 2,
+                byte_end: 5,
+                text_bytes: 9,
+            })
+        );
+        // The neighbour of all three: the unmutated spans still decode.
+        assert_eq!(reconstruct(len, &digest, &[a, b]).as_deref(), Ok("ab\ncd"));
+    }
+
+    #[test]
+    fn a_cover_that_rebuilds_the_wrong_bytes_is_named_by_its_digest() {
+        let (len, _, [a, b]) = covered();
+        let other = ContentDigest::of(b"ab\ncD");
+        assert_eq!(
+            reconstruct(len, &other, &[a, b]),
+            Err(ReconstructError::DigestMismatch {
+                expected: other.to_hex(),
+                reconstructed: ContentDigest::of(b"ab\ncd").to_hex(),
+            })
+        );
+        // The neighbour: the digest the graph would actually state.
+        let stated = ContentDigest::of(b"ab\ncd");
+        assert!(reconstruct(len, &stated, &[a, b]).is_ok());
+    }
+
+    #[test]
+    fn an_empty_document_is_a_lawful_graph_of_nothing() {
+        assert_eq!(
+            reconstruct(0, &ContentDigest::of(b""), &[]).as_deref(),
+            Ok("")
+        );
+        // The neighbour in the other direction: a declared length with no
+        // spans at all is one uncovered run.
+        assert_eq!(
+            reconstruct(1, &ContentDigest::of(b"x"), &[]),
+            Err(ReconstructError::UncoveredRange {
+                byte_start: 0,
+                byte_end: 1,
+            })
+        );
+    }
+}
